+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.4.10
 - Improved performance of network scanning (raoulvdberge)
+- Fixed crash when attempting to get direction of a node (raoulvdberge)
 
 ### 1.4.9
 - Fixed bug where inventory data was lost sometimes upon opening the world (raoulvdberge)